requirements.txt
streamlit 
youtube_transcript_api
google-generativeai
python-dotenv
pathlib
gtts
SpeechRecognition
textblob